Better (less) wrapping of proper lists.
This change represents an intention to store Markdown lists in the
canonical Markdown format, with single line breaks from point to point,
as opposed to double line breaks, which Markdown can parse but with
inferior HTML results.

The solution is not perfect. A special subsystem would be needed to
automatically (un)wrap list items effectively.