API break
Rebuilt the package on top of `punwrap`, removing the homegrown regex
logic and the passing of wordsplitters as parameters (this is the API
break), which `punwrap` cannot handle.

Results will be different.

At the time of this commit, this rebuild is known to cause some bad
behaviour; cf. https://github.com/mgeisler/textwrap/issues/384.